@@ -6,13 +6,13 @@
         { "name": "Christian Högl", "email": "chrigu@sirprize.me" }
     ],
     "require": {
-        "php": ">=5.4.0",
+        "php": "^7.3",
         "symfony/event-dispatcher": "~2.8|~3.0|~4.0",
         "symfony/console": "~2.8|~3.0|~4.0",
         "monolog/monolog": "~1.12"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": { "Kompakt\\DirectoryRunner\\": "lib/" }
